pyproject.toml
[tool.poetry]
name = "screenlogic-ui-py"
version = "0.1.0"
description = ""
authors = ["Michael Mucciarone <mikemucc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "*"
dash = "*"
dash-bootstrap-components = {extras = ["pandas"], version = "^1.0.2"}
requests = "*"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pylint = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
